requirements.txt
BeautifulSoup==3.2.1
Cython==0.23.4
Django==1.5.12
Pillow==2.3.0
PyYAML==3.10
cGPolyEncode==0.1.1
django-constance==1.0.1
django-json-rpc==0.7.2
django-picklefield==0.3.2
django-redis==4.3.0
django-simple-captcha==0.5.1
gtfs-realtime-bindings==0.0.6
ipython==4.1.2
lxml==3.3.3
marisa-trie==0.7.2
Markdown
paramiko==1.16.0
protobuf==3.11.2
psycopg2-binary==2.8.6
ptyprocess==0.5.1
pycha==0.7.0
pycrypto==2.6.1
pycurl==7.19.3
pyproj==1.9.5.1
pyshp==1.1.4
pytz==2015.7
redis==2.10.5
requests==2.9.1
rpyc==3.3.0

